# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavelength API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/discovery": {
      "get": {
        "summary": "Compatibility-ranked candidates for the signed-in user.",
        "tags": [
          "discovery"
        ],
        "parameters": [
          {
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "integer"
            },
            "in": "query",
            "name": "minScore",
            "required": false
          },
          {
            "schema": {
              "minimum": 18,
              "maximum": 120,
              "type": "integer"
            },
            "in": "query",
            "name": "minAge",
            "required": false
          },
          {
            "schema": {
              "minimum": 18,
              "maximum": 120,
              "type": "integer"
            },
            "in": "query",
            "name": "maxAge",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "cards"
                  ],
                  "properties": {
                    "cards": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "user",
                          "overallScore",
                          "music",
                          "movie",
                          "book",
                          "sharedHighlights",
                          "divergenceHighlights",
                          "tasteDna",
                          "yourTasteDna",
                          "narrative"
                        ],
                        "properties": {
                          "user": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "age",
                              "gender",
                              "city",
                              "bio",
                              "photoUrl"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "age": {
                                "anyOf": [
                                  {
                                    "type": "integer"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "gender": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "city": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "bio": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "photoUrl": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            }
                          },
                          "overallScore": {
                            "type": "integer"
                          },
                          "music": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "movie": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "book": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "sharedHighlights": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "domain",
                                "label"
                              ],
                              "properties": {
                                "domain": {
                                  "type": "string"
                                },
                                "label": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "divergenceHighlights": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "required": [
                                "domain",
                                "label",
                                "heldBy"
                              ],
                              "properties": {
                                "domain": {
                                  "type": "string"
                                },
                                "label": {
                                  "type": "string"
                                },
                                "heldBy": {
                                  "type": "string"
                                }
                              }
                            }
                          },
                          "tasteDna": {
                            "type": "object",
                            "required": [
                              "niche",
                              "melancholic",
                              "contemporary",
                              "maximalist",
                              "challenging"
                            ],
                            "properties": {
                              "niche": {
                                "type": "number"
                              },
                              "melancholic": {
                                "type": "number"
                              },
                              "contemporary": {
                                "type": "number"
                              },
                              "maximalist": {
                                "type": "number"
                              },
                              "challenging": {
                                "type": "number"
                              }
                            }
                          },
                          "yourTasteDna": {
                            "type": "object",
                            "required": [
                              "niche",
                              "melancholic",
                              "contemporary",
                              "maximalist",
                              "challenging"
                            ],
                            "properties": {
                              "niche": {
                                "type": "number"
                              },
                              "melancholic": {
                                "type": "number"
                              },
                              "contemporary": {
                                "type": "number"
                              },
                              "maximalist": {
                                "type": "number"
                              },
                              "challenging": {
                                "type": "number"
                              }
                            }
                          },
                          "narrative": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/discovery/{userId}/narrative": {
      "get": {
        "summary": "The AI compatibility narrative for one pairing (spec 3.5).",
        "tags": [
          "discovery"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "narrative"
                  ],
                  "properties": {
                    "narrative": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/compatibility/{userId}": {
      "get": {
        "summary": "Full compatibility breakdown for one pairing.",
        "tags": [
          "discovery"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user",
                    "overallScore",
                    "music",
                    "movie",
                    "book",
                    "sharedHighlights",
                    "divergenceHighlights",
                    "tasteDna",
                    "yourTasteDna"
                  ],
                  "properties": {
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "age",
                        "gender",
                        "city",
                        "bio",
                        "photoUrl"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "age": {
                          "anyOf": [
                            {
                              "type": "integer"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "gender": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "city": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "bio": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "photoUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    },
                    "overallScore": {
                      "type": "integer"
                    },
                    "music": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "movie": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "book": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "sharedHighlights": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "domain",
                          "label"
                        ],
                        "properties": {
                          "domain": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "divergenceHighlights": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "domain",
                          "label",
                          "heldBy"
                        ],
                        "properties": {
                          "domain": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "heldBy": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "tasteDna": {
                      "type": "object",
                      "required": [
                        "niche",
                        "melancholic",
                        "contemporary",
                        "maximalist",
                        "challenging"
                      ],
                      "properties": {
                        "niche": {
                          "type": "number"
                        },
                        "melancholic": {
                          "type": "number"
                        },
                        "contemporary": {
                          "type": "number"
                        },
                        "maximalist": {
                          "type": "number"
                        },
                        "challenging": {
                          "type": "number"
                        }
                      }
                    },
                    "yourTasteDna": {
                      "type": "object",
                      "required": [
                        "niche",
                        "melancholic",
                        "contemporary",
                        "maximalist",
                        "challenging"
                      ],
                      "properties": {
                        "niche": {
                          "type": "number"
                        },
                        "melancholic": {
                          "type": "number"
                        },
                        "contemporary": {
                          "type": "number"
                        },
                        "maximalist": {
                          "type": "number"
                        },
                        "challenging": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/swipe": {
      "post": {
        "summary": "Like or pass on a candidate. A mutual like creates a match.",
        "tags": [
          "discovery"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetId",
                  "direction"
                ],
                "properties": {
                  "targetId": {
                    "type": "string"
                  },
                  "direction": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "like"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "pass"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "matched",
                    "matchId",
                    "requested"
                  ],
                  "properties": {
                    "matched": {
                      "type": "boolean"
                    },
                    "matchId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "requested": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/matches/{matchId}/unmatch": {
      "post": {
        "tags": [
          "matches"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "matchId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{userId}/block": {
      "post": {
        "tags": [
          "safety"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "userId",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}